// src/interfaces/FileType.ts
enum FileType {
    USER_PROFILE_PICTURE = "USER_PROFILE_PICTURE",
    USER_COVER_PICTURE = "USER_COVER_PICTURE",
    ALBUM_COVER = "ALBUM_COVER",
    PLAYLIST_COVER = "PLAYLIST_COVER",
    MUSIC_COVER = "MUSIC_COVER",
    MUSIC_AUDIO = "MUSIC_AUDIO",
    CLIP_THUMBNAIL = "CLIP_THUMBNAIL",
    CLIP_VIDEO = "CLIP_VIDEO",
}

export default FileType;